Before:

```
BenchmarkItems-4                                 200       8286514 ns/op
BenchmarkMarshalJson-4                            50      25730621 ns/op
BenchmarkStrconv-4                          20000000           125 ns/op
BenchmarkSingleInsertAbsent-4                1000000          1600 ns/op
BenchmarkSingleInsertPresent-4               5000000           259 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4       500000          4885 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4      500000          4669 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4      500000          4608 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4     300000          5480 ns/op
BenchmarkMultiInsertSame-4                    500000          2738 ns/op
BenchmarkMultiGetSame-4                      1000000          1032 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4       300000          5963 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4      300000          5792 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4      300000          5737 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4     300000          5504 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4           300000          4404 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4          300000          4623 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4          300000          4402 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4         300000          4104 ns/op
BenchmarkKeys-4                                  500       3339432 ns/op
ok      _/home/rossmohax/repos/external/concurrent-map  37.088s
```

After

```
BenchmarkItems-4                                 200       8357663 ns/op
BenchmarkMarshalJson-4                            50      26142281 ns/op
BenchmarkStrconv-4                          20000000           127 ns/op
BenchmarkSingleInsertAbsent-4                1000000          1890 ns/op
BenchmarkSingleInsertPresent-4               5000000           257 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4       500000          4666 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4      500000          4447 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4      500000          4881 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4     300000          5393 ns/op
BenchmarkMultiInsertSame-4                    500000          2633 ns/op
BenchmarkMultiGetSame-4                      2000000           991 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4       300000          5495 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4      300000          5788 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4      300000          5528 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4     300000          5423 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4           300000          4322 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4          300000          4422 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4          500000          4154 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4         500000          4040 ns/op
BenchmarkKeys-4                                  500       3329185 ns/op
ok      _/home/rossmohax/repos/external/concurrent-map  40.550s
```