@orcaman 2016-12-05
@redbaron thanks for your contribution. I'm wondering, what is gained by using pointer receivers for the functions? as your benchmark shows there is no performance improvement.